Move LCD keyframes to its own file
