+++ keyboards/ergodox/keymaps/default/visualizer.c
@@ -27,6 +27,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #endif
 
 #include "visualizer.h"
+#include "lcd_keyframes.h"
 #include "system/serial_link.h"
 #include "led.h"
 
@@ -122,7 +123,7 @@ static keyframe_animation_t lcd_layer_display = {
     .num_frames = 1,
     .loop = false,
     .frame_lengths = {gfxMillisecondsToTicks(0)},
-    .frame_functions = {keyframe_display_layer_and_led_states}
+    .frame_functions = {lcd_keyframe_display_layer_and_led_states}
 };
 
 static keyframe_animation_t suspend_animation = {
@@ -130,7 +131,7 @@ static keyframe_animation_t suspend_animation = {
     .loop = false,
     .frame_lengths = {0, gfxMillisecondsToTicks(1000), 0},
     .frame_functions = {
-            keyframe_display_layer_text,
+            lcd_keyframe_display_layer_text,
             keyframe_animate_backlight_color,
             keyframe_disable_lcd_and_backlight,
     },
