+++ quantum/visualizer/visualizer.mk
@@ -34,6 +34,7 @@ endif
 
 ifeq ($(strip $(LCD_ENABLE)), yes)
 SRC += $(VISUALIZER_DIR)/lcd_backlight.c
+SRC += $(VISUALIZER_DIR)/lcd_keyframes.c
 OPT_DEFS += -DLCD_BACKLIGHT_ENABLE
 endif
